% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{get_mean_clusters}
\alias{get_mean_clusters}
\title{Fit Kmeans on globs}
\usage{
get_mean_clusters(data, fit_name, n_clust, iter, n_start, seed = NULL)
}
\arguments{
\item{data}{Dataframe of 2-d space to fit on.}

\item{fit_name}{Name of kmeans fit}

\item{n_clust}{An integer number of clusters to fit.}

\item{iter}{An integer number of max iterations for the kmeans algorithm to converge to.}

\item{n_start}{An integer number of max iterations for the kmeans algorithm to converge to}

\item{seed}{Random seed to use to make the results reproducible.}
}
\value{
Dataframe of covariates, predictions, and title of fit.
}
\description{
Fit Kmeans on globs
}
