% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kprototypes.R
\name{get_proto_clusters}
\alias{get_proto_clusters}
\title{Fit Kprototypes on globs}
\usage{
get_proto_clusters(data, fit_name, n_clust, iter, n_start, seed = NULL)
}
\arguments{
\item{data}{Dataframe of a 3-d space to fit on. Two continuous dimensions and one categorical dimension.}

\item{fit_name}{Name of kprototypes fit}

\item{n_clust}{An integer number of clusters to fit.}

\item{iter}{An integer number of max iterations for the kprototypes algorithm to converge to.}

\item{n_start}{An integer number of max iterations for the kprototypes algorithm to converge to}

\item{seed}{Random seed to use to make the results reproducible.}
}
\value{
Dataframe of covariates, predictions, and title of fit.
}
\description{
Fit Kprototypes on globs
}
