% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clusters.R
\name{make_circle}
\alias{make_circle}
\title{Create a 2-d circular Boolean mask}
\usage{
make_circle(center, r, pts)
}
\arguments{
\item{center}{Cartesian center of the blob specified as c(x, y).}

\item{r}{Radius of the blob.}

\item{pts}{Dataframe of Cartesian point coordinates to draw from.}
}
\value{
Boolean mask when applied forms a circle from a point cloud.
}
\description{
Create a 2-d circular Boolean mask
}
