% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clusters.R
\name{make_cluster}
\alias{make_cluster}
\title{Extract a circular cluster from a square point field}
\usage{
make_cluster(center, r, n_pts)
}
\arguments{
\item{center}{Cartesian center of the blob specified as c(x, y).}

\item{r}{Radius of the blob.}

\item{n_pts}{Number of points in the cluster.}
}
\value{
Vector of points in the shape of a circle with radius r and center (x,y).
}
\description{
Extract a circular cluster from a square point field
}
