% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clusters.R
\name{make_clusters}
\alias{make_clusters}
\title{Make circular clusters}
\usage{
make_clusters(
  n_clust,
  centers = "random",
  c_range,
  radii = c(1, 5),
  density,
  categorical = FALSE,
  seed = NULL
)
}
\arguments{
\item{n_clust}{An integer number of clusters to generate.}

\item{centers}{If "random" then centers are randomly generated from the c_range sequence. Otherwise, a list of cartesian coordinates for cluster centers should be provided.}

\item{c_range}{Square Cartesian space from which centers can be drawn. Expressed as a vector of two integers.}

\item{radii}{A vector of two integers, the range to choose the radius of each cluster from.}

\item{density}{The relative number of points per cluster. A function of the cluster radius and beta distribution.}

\item{categorical}{A fractional value if provided. Determines how useful a categorical variable should be with 1 being the most useful.}

\item{seed}{Random seed to use to make the results reproducible.}
}
\value{
Dataframe, a set of labeled 2-d clusters.
}
\description{
Make circular clusters
}
