% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clusters.R
\name{make_circle}
\alias{make_circle}
\title{Add together two numbers}
\usage{
make_circle(center, r, pts)
}
\arguments{
\item{center}{Cartesian center of the blob specified as c(x, y).}

\item{r}{Radius of the blob.}

\item{pts}{Dataframe of cartesian point coordinates to draw from.}
}
\value{
Vector of points in the shape of a circle with radius r and center (x,y).
}
\description{
Add together two numbers
}
